requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
Flask
ipykernel
logger
pygithub 
# -e .